feat: added CopyButton components
