+++ src/components/CopyButton/index.tsx
@@ -1,0 +1,2 @@
+export { default as CopyButton } from './CopyButton';
+export type { CopyButtonProps } from './types';
