+++ src/components/CopyButton/types.ts
@@ -1,0 +1,3 @@
+export interface CopyButtonProps {
+    text?: string | number;
+}
