+++ src/components/Typography/Typography.styles.ts
@@ -1,8 +1,7 @@
-import styled, { css } from 'styled-components';
+import { css } from 'styled-components';
 import color from '../../styles/colors';
 import type { TypographyProps, variantType, weightType } from './types';
 import './fonts/fonts.css';
-import { resetButtonCSS } from '../../styles/reset';
 
 const openSans = css`
     @import url('//fonts.googleapis.com/css?family=Open+Sans');
@@ -160,25 +159,6 @@ export const getFontStyle = (variant: variantType) => {
             return body16;
     }
 };
-
-export const CopyIconStyled = styled.button`
-    ${resetButtonCSS}
-    align-items: center;
-    display: inline-flex;
-    margin-left: 0.1em;
-    position: absolute;
-
-    & :first-child {
-        height: 1em;
-        transition: fill 0.2s ease-out;
-        fill: ${color.blueSkyDark};
-        width: 1em;
-    }
-
-    &:hover > svg {
-        fill: ${color.blue};
-    }
-`;
 
 export const getTypographyStyle = (p: TypographyProps) => css`
     ${getFontStyle(p.variant || 'body16')};
